EOS-14134: Move sw_udpate logic back to __init__.py file. Postpone this for the separate PR
